contact/requirements.txt
biopython==1.81
certifi==2023.7.22
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1666183775483/work
charset-normalizer==3.3.1
dgl==0.6.1
docopt==0.6.2
future @ file:///home/conda/feedstock_root/build_artifacts/future_1649010143709/work
h5py==3.8.0
idna==3.4
networkx==2.6.3
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1649806299270/work
psutil==5.9.6
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
requests==2.31.0
scipy==1.7.3
torch @ file:///home/conda/feedstock_root/build_artifacts/pytorch-recipe_1640879027892/work
tqdm==4.66.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1688315532570/work
urllib3==2.0.7
yarg==0.1.9
